# server/internal/handlers/docs/swagger.yaml
basePath: /
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.createAccountResponse:
    properties:
      id:
        type: string
    type: object
...
  handlers.statsResponse:
    properties:
      data:
        $ref: '#/definitions/models.DashboardStats'
    type: object
  models.Account:
    properties:
      api_key:
        type: string
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
...
  models.AlertStats:
    properties:
      critical:
        type: integer
      total:
        type: integer
    type: object
  models.CreateAccount:
    properties:
      api_key:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  models.DashboardStats:
    properties:
      activeAlerts:
        $ref: '#/definitions/models.AlertStats'
      authFail:
        $ref: '#/definitions/models.Trend'
      errorRate:
        $ref: '#/definitions/models.Trend'
      requests:
        $ref: '#/definitions/models.Trend'
    type: object
...
  models.Log:
    properties:
      agentId:
        type: string
      event:
        type: string
      id:
        type: string
      level:
        type: string
      logTime:
        type: string
      message:
        type: string
      recordedAt:
        type: string
      userId:
        type: string
    type: object
  models.LogVolumeStats:
    properties:
      app_logs:
        type: integer
      error_count:
        type: integer
      has_error:
        type: boolean
      nginx_logs:
        type: integer
      time:
        type: string
    type: object
...
  models.NginxLog:
    properties:
      agentId:
        type: string
      bytes:
        type: integer
      id:
        type: string
      ipaddress:
        type: string
      logTime:
        type: string
      method:
        type: string
      path:
        type: string
      recordedAt:
        type: string
      status:
        type: integer
      userAgent:
        type: string
    type: object
...
  models.Trend:
    properties:
      percent:
        type: number
      value:
        type: number
    type: object
  models.UpdateAccount:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API server for application
  title: Sentinel API
  version: "1.0"
paths:
  /sentinel/accounts:
    post:
      consumes:
      - application/json
      description: Create Account
      parameters:
      - description: Create Account
        in: body
        name: create
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.createAccountResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Account
      tags:
      - account
  /sentinel/accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Account
      parameters:
      - description: account id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Account
      tags:
      - account
    get:
      consumes:
      - application/json
      description: Get Account by ID
      parameters:
      - description: account id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Account
      tags:
      - account
    put:
      consumes:
      - application/json
      description: Update Account
      parameters:
      - description: account id
        in: path
        name: id
        required: true
        type: string
      - description: update account
        in: body
        name: update
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Account
      tags:
      - account
...
  /sentinel/applogs:
    get:
      description: Get app logs by filter
      parameters:
      - description: agent id
        in: query
        name: agent_id
        type: string
      - description: user id
        in: query
        name: user_id
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: level
        in: query
        name: level
        type: string
      - description: filter from recorded_at (RFC3339 format)
        in: query
        name: from
        type: string
      - description: Filter to recorded_at (RFC3339 format)
        in: query
        name: to
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get app logs by filter
      tags:
      - logs
  /sentinel/applogs/{id}:
    get:
      consumes:
      - application/json
      description: Get applog by id
      parameters:
      - description: app log id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Log'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get applog by id
      tags:
      - logs
...
  /sentinel/logvolume:
    get:
      consumes:
      - application/json
      description: Get log volume stats
      parameters:
      - description: agent id
        in: query
        name: agent_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.LogVolumeStats'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get log volume statistics for an agent
      tags:
      - dashboard
  /sentinel/me:
    get:
      consumes:
      - application/json
      description: Get Me
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current account
      tags:
      - account
...
  /sentinel/nginxlogs:
    get:
      consumes:
      - application/json
      description: get nginxlogs by filter
      parameters:
      - description: agent id
        in: query
        name: agent_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: filter from recorded_at (RFC3339 format)
        in: query
        name: from
        type: string
      - description: Filter to recorded_at (RFC3339 format)
        in: query
        name: to
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: get nginxlogs by filter
      tags:
      - logs
  /sentinel/nginxlogs/{id}:
    get:
      consumes:
      - application/json
      description: Get nginx log by id
      parameters:
      - description: nginx log id
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NginxLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get nginx log by id
      tags:
      - logs
...
  /sentinel/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard stats
      parameters:
      - description: agent id
        in: query
        name: agent_id
        type: string
      - default: 1h
        description: 'time period: 1h, 24h, 7d'
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.statsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get dashboard statistics
      tags:
      - dashboard
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"